roll back contextType enum to odo.ts
