+++ src/odo.ts
@@ -18,7 +18,7 @@ import yaml = require('js-yaml');
 import fs = require('fs');
 import * as odo from './odo/config';
 import { ComponentSettings } from './odo/config';
-import { GlyphChars, ContextType } from './util/constants';
+import { GlyphChars } from './util/constants';
 
 const Collapsed = TreeItemCollapsibleState.Collapsed;
 
@@ -31,6 +31,20 @@ export interface OpenShiftObject extends QuickPickItem {
     contextPath?: Uri;
     deployed: boolean;
     path?: string;
+}
+
+export enum ContextType {
+    CLUSTER = 'cluster',
+    PROJECT = 'project',
+    APPLICATION = 'application',
+    COMPONENT = 'component_not_pushed',
+    COMPONENT_PUSHED = 'component',
+    COMPONENT_NO_CONTEXT = 'component_no_context',
+    SERVICE = 'service',
+    STORAGE = 'storage',
+    CLUSTER_DOWN = 'cluster_down',
+    LOGIN_REQUIRED = 'login_required',
+    COMPONENT_ROUTE = 'component_route'
 }
 
 function verbose(_target: any, key: string, descriptor: any) {
