+++ src/util/constants.ts
@@ -5,20 +5,6 @@
 
 'use strict';
 
-export enum ContextType {
-    CLUSTER = 'cluster',
-    PROJECT = 'project',
-    APPLICATION = 'application',
-    COMPONENT = 'component_not_pushed',
-    COMPONENT_PUSHED = 'component',
-    COMPONENT_NO_CONTEXT = 'component_no_context',
-    SERVICE = 'service',
-    STORAGE = 'storage',
-    CLUSTER_DOWN = 'cluster_down',
-    LOGIN_REQUIRED = 'login_required',
-    COMPONENT_ROUTE = 'component_route'
-}
-
 export enum GlyphChars {
     Asterisk = '\u2217',
     Check = '\u2713',
